Move epoch math to math.rs
